requirements.txt
ts3
pymysql
reportTelegram
teamSpeakTelegram
